Properly handling async model for initializing MediaCapture
The documentation states that in apps that use C# or C++, the first use of the MediaCapture object to call InitializeAsync family methods should be on the STA thread. Calls from an MTA thread may result in undefined behavior.

It seems that, in UWP, calling MediaCapture::InitializeAsync from a MTA thread works, but calling MediaCapture::InitializeWithSettingsAsync from a MTA thread doesn’t work properly. This restriction seems to not repro in Win32.

This change transfers the MediaCapture initialization call to the STA thread (UI thread). Additionally, this change handles correctly the async pattern used for requesting camera and microphone permissions from the user.