Checking for async operation finalizations before waiting for its finalization
Async operations might be finalized before the code for setting up the wait in the calling thread have chance to run. This change checks if the async operation has been completed before waiting for it.
Additionally, this change also fixes the bug where the UWP code wasn’t being used for waiting for user input for granting camera and mic access.